Module Command.Make

Parameters

module IO : Conn.IO

module Codec : Conn.Codec.S

module Store : Irmin.Generic_key.S

Signature

module Conn : Conn.S with module IO = IO

module Store = Store

Irmin Store type

module Tree : Tree.S with type kinded_key = Store.Tree.kinded_key and type concrete = Store.Tree.concrete with type kinded_key = Store.Tree.kinded_key with type concrete = Store.Tree.concrete

Tree wraps Store.Tree to avoid encoding/decoding trees more than needed

module Commit : sig ... end

Commits

module Server_info : sig ... end

Used to track information about the server

type context = {

conn : Conn.t;

config : Irmin.Backend.Conf.t;

repo : Store.Repo.t;

mutable watch : Store.watch option;

mutable branch_watch : Store.Backend.Branch.watch option;

}

context is passed to every command as the first argument

module type CMD = sig ... end

type t = (module CMD)

Command type

val name : t -> string

Get the name of a command

val of_name : string -> t

Find a command by name, the name should be lowercase

val commands : (string * t) list

A list of all registered commands

module Commands : sig ... end