Module Commands.Store

type t = store

val t : t Irmin.Type.t

type write_options = (bool option * int option) * (bool option * Store.hash list option)

val write_options_t : write_options Irmin.Type.t

module Find : CMD with type req = t * Store.path and type res = Store.contents option

Find a value in the store

module Remove : CMD with type req = write_options * (t * Store.path) * Store.Info.t and type res = unit

Remove a value from the store

module Find_tree : CMD with type req = t * Store.path and type res = Store.Tree.concrete option

Get a tree from the store

module Mem : CMD with type req = t * Store.path and type res = bool

Check for the existence of a value in the store

module Mem_tree : CMD with type req = t * Store.path and type res = bool

Check for the existence of a tree in the store